Skipped 4.22-4.23 due to cofusion + time

+++ controllers/blogs.js
@@ -2,26 +2,19 @@ const blogRouter = require('express').Router();
 const Blog = require("../models/blog");
 const User = require('../models/user')
 const jwt = require("jsonwebtoken");
+const middleware = require("../utils/middleware")
 
-const getTokenFrom = (request) => {
-  const authorization = request.get("authorization");
-  if(authorization && authorization.startsWith("Bearer ")) {
-    return authorization.replace("Bearer ", "")
-  }
-  return null
-}
+
 
 blogRouter.get('/', async (request, response) => {
     const blogs = await Blog.find({}).populate("user", { username: 1, name: 1 })
       response.json(blogs)
   })
 
-blogRouter.post('/', async (request, response) => {
+blogRouter.post('/', middleware.userExtractor, async (request, response) => {
   const body = request.body;
   const decodetoken = jwt.verify(request.token, process.env.SECRET)
-  if(!decodetoken.id) {
-    return response.status(401).json({error: "token invalid"})
-  }
+  
   const user = await User.findById(decodetoken.id)
 
   const blog = new Blog ({
@@ -43,23 +36,40 @@ blogRouter.post('/', async (request, response) => {
   }
 })
 
-blogRouter.delete(`/:id`, async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end();
+blogRouter.delete(`/:id`, middleware.userExtractor, async (request, response) => { 
+  const blog = await Blog.findById(request.params.id)
+  const decodetoken = jwt.verify(request.token, process.env.SECRET)
+  console.log( `${blog.user} || ${decodetoken.id}`)
+  if(blog.user.toString() === decodetoken.id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end();
+ } else if(blog.user.toString() !== decodetoken.id.toString()){ 
+    return response.status(403).json({ error: "Unauthorized, incorrect authorization" })
+ } else {
+    return response.status(401).json({ error: "Authentication is required to delete" })
+ }
+  
 })
 
-blogRouter.put(`/:id`, async (request, response) => {
+blogRouter.put(`/:id`, middleware.userExtractor, async (request, response) => {
+  const blog = await Blog.findById(request.params.id)
   const body = request.body;
+  const decodetoken = jwt.verify(request.token, process.env.SECRET)
 
-  const replaceBlog = {
-    title: body.title,
-    author: body.author,
-    url: body.url,
-    likes: body.likes || 0
+  
+  if(blog.user.toString() === decodetoken.id.toString()) {
+    const replaceBlog = {
+      title: body.title,
+      author: body.author,
+      url: body.url,
+      likes: body.likes || 0
+    }
+
+    const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, replaceBlog, { new: true })
+    response.json(updatedBlog)
+  } else {
+    return response.status(403).json({ error: "Unauthorized, incorrect authorization" })
   }
-
-  await Blog.findByIdAndUpdate(request.params.id, replaceBlog, { new: true })
-    response.json(replaceBlog)
 })
 
 module.exports = blogRouter;
